periodics.pde	better logic for fix_global_next() call in series.
This version calculates global next event[s] *once* when the next
event could possibly have changed. See fix_global_next().

If a series of tasks are sceduled for the same time they will
be called in fast sequence *without* fix_global_next() in between them.
After all task with the same time are done fix_global_next() is called.

If a task sets up another task with the same next it *can* decide to do
the fix itself, if it seems appropriate. Normally this is not required.