'Continuous' display implemented in the sketch.
After doing a lot of work on implementing 'continuous' display in Menu
I realized, that pulses does not need this rather obscure functionality
in the Menu library. Only example/softboard uses it, so let me define
it there.

Later on this type of background tasks are planed to be handled by
the Pulses library anyway.

So I decided to throw away my work on 'continuous display' type
functionality in the Menu.